#46 [MOVE] : model news 패키지 위치 이동
