+++ core/model/src/main/java/com/teamwable/model/news/NewsMatchScoreModel.kt
@@ -1,4 +1,4 @@
-package com.teamwable.model
+package com.teamwable.model.news
 
 data class NewsMatchScoreModel(
     val gameDate: String,
